requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
pillow==10.4.0
pytils==0.4.1
ipython==8.28.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
coverage==7.6.3

